provider: Skip provider init on silent auth and offline
It doesn't make any sense to initialize the
provider if we're trying to do a silent auth
(which requires prior data stored in the existing
provider), or if we're offline (we can't init a
provider when we're offline). These appear to be
causing the daemon to freeze when we're in a
systemd sandbox.

Signed-off-by: David Mulder <dmulder@samba.org>