Windows Hello PIN implementation
This causes Himmelblau to enroll the
authenticating user in Windows Hello for Pin
authentication. It uses the msal rust library
to associate the user chosen Pin with a rsa key,
which is then enrolled in Windows Hello.
This eliminates the problem with MFA prompts at
the lock screen (which didn't work). The user now
treats the Pin code as a local password.

Signed-off-by: David Mulder <dmulder@samba.org>