Always use the spn of the user for nss requests
This is the cause of the failing whoami requests,
because it provides the uid of the user for the
request. As long as the cached user token hadn't
expired, the resolver would respond. But when it
needed refreshed, Himmelblau couldn't even find
the id provider, because it was trying to match
the uid against a domain name on the spn.

Signed-off-by: David Mulder <dmulder@samba.org>