Implement Hello Pin changes via PAM
This implements Pin changes via pam (a user will
call the terminal `passwd` command). This will
require more interaction than a typical call to
`passwd`, since we need to authenitcate to Entra
Id to register the new Pin. We could use either
an existing Hello Key or the PRT to register the
key, but not if we authenticated using a DAG (or
SFA, for that matter). This extra authentication
step forces the token to have an ngcmfa amr (if
using the interactive flow).

Signed-off-by: David Mulder <dmulder@samba.org>