Implement Hello Pin changes via PAM
This implements Pin changes via pam (a user will
call the terminal `passwd` command). This will
require more interaction than a typical call to
`passwd`, since we need to authenticate to Entra
Id to register the new Pin. This extra
authentication step forces the token to have an
ngcmfa amr.

Signed-off-by: David Mulder <dmulder@samba.org>