Entra Id no longer permits SFA enrollment
Entra Id is now rejecting authentication attempts
to the enrollment service with Single Factor
Authentication. We can work around this by
permitting users to authenticate with SFA if the
host is already enrolled. This means a user with
MFA will be required to enroll the host (which
just requires authenticating once).

Signed-off-by: David Mulder <dmulder@samba.org>