Implement `decoupled hello` behavior
This implements the behavior observed in Authd,
where a 'local password' is possible even when
authentication is performed via a DAG. Unlike
Authd, this 'local password' (we just treat it
like a Linux Hello PIN here) is TPM bound, and
the PIN unlocks a `decoupled hello` key. This
key is not actually bound to Entra Id, but is
instead a local only PIN. The user MUST possess
the PIN in order to unlock the key, and only that
unlocked key can then decrypt a cached PRT. This
key is both PIN bound AND host bound, so only
that user can unlock the key on that host. This
permits us to emulate 'local password' behavior,
as well as provide SSO via a PRT.

Signed-off-by: David Mulder <dmulder@samba.org>