Clarify `domain` SHOULD match login domain
The MUST used here was misleading. Technically
it can be ANY domain for the tenant, but using the
domain which users will authenticate with is less
prone to errors. For example, if they enroll in
an alias domain, then authenticate with a
different domain, they will experience transient
authentication errors from Entra Id (this is NOT
a Himmelblau bug, but an Entra Id bug)!

Signed-off-by: David Mulder <dmulder@samba.org>