systemd: install files to /usr/lib/, not /etc/
/etc/ is for local configurations by the users/admins. Packages must
install to /usr/lib/, so that users can override them locally if
they want to do so.
Packages need to be resilient against factory resets (deleting /etc/
and /var/ and rebooting).

Also add a more descriptive, namespaced name for the gdm3 override,
as override.conf is too generic and could conflict.

Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>